business layer created

+++ src/main/java/dev/patika/VeterinaryManagementSystem/business/abstracts/IAnimalService.java
@@ -6,11 +6,11 @@ import org.springframework.data.domain.Page;
 public interface IAnimalService {
     Animal save(Animal animal);
 
-    Animal get(int id);
+    Animal get(long id);
 
     Animal update(Animal animal);
 
-    boolean delete(int id);
+    boolean delete(long id);
 
     Page<Animal> cursor(int page, int pageSize);
 }
