+++ src/main/java/dev/patika/VeterinaryManagementSystem/business/abstracts/IAppointmentService.java
@@ -6,11 +6,11 @@ import org.springframework.data.domain.Page;
 public interface IAppointmentService {
     Appointment save(Appointment appointment);
 
-    Appointment get(int id);
+    Appointment get(long id);
 
     Appointment update(Appointment appointment);
 
-    boolean delete(int id);
+    boolean delete(long id);
 
     Page<Appointment> cursor(int page, int pageSize);
 }
