+++ src/main/java/dev/patika/VeterinaryManagementSystem/business/abstracts/IDoctorService.java
@@ -6,11 +6,11 @@ import org.springframework.data.domain.Page;
 public interface IDoctorService {
     Doctor save(Doctor doctor);
 
-    Doctor get(int id);
+    Doctor get(long id);
 
     Doctor update(Doctor doctor);
 
-    boolean delete(int id);
+    boolean delete(long id);
 
     Page<Doctor> cursor(int page, int pageSize);
 }
