+++ src/main/java/dev/patika/VeterinaryManagementSystem/business/abstracts/IVaccineService.java
@@ -6,11 +6,11 @@ import org.springframework.data.domain.Page;
 public interface IVaccineService {
     Vaccine save(Vaccine vaccine);
 
-    Vaccine get(int id);
+    Vaccine get(long id);
 
     Vaccine update(Vaccine vaccine);
 
-    boolean delete(int id);
+    boolean delete(long id);
 
     Page<Vaccine> cursor(int page, int pageSize);
 }
